+++ src/Params/EOptionsParams.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ZekyWolf\LGSQ\Params;
+
+abstract class EOptionsParams
+{
+    public const CURL_CONNECT_TIMEOUT = 'curl_connect_timeout';
+    public const CURL_TIMEOUT = 'curl_timeout';
+}