+++ src/Protocols/Query06.php
@@ -2,12 +2,12 @@
 
 namespace ZekyWolf\LGSQ\Protocols;
 
-use ZekyWolf\LGSQ\Helpers\{
-    Parse\Byte,
-    Parse\Colors,
-    Parse\ParseString,
-    EServerParams as SParams,
-    EConnectionParams as CParams
+use ZekyWolf\LGSQ\{
+    Helpers\Parse\Byte,
+    Helpers\Parse\Colors,
+    Helpers\Parse\ParseString,
+    Params\EServerParams as SParams,
+    Params\EConnectionParams as CParams
 };
 
 class Query06