+++ src/Protocols/Query07.php
@@ -2,10 +2,10 @@
 
 namespace ZekyWolf\LGSQ\Protocols;
 
-use ZekyWolf\LGSQ\Helpers\{
-    Parse\Colors,
-    EServerParams as SParams,
-    EConnectionParams as CParams
+use ZekyWolf\LGSQ\{
+    Helpers\Parse\Colors,
+    Params\EServerParams as SParams,
+    Params\EConnectionParams as CParams
 };
 
 class Query07